+++ near-sdk/src/collections/legacy_tree_map.rs
@@ -6,6 +6,7 @@ use std::ops::Bound;
 
 use crate::collections::UnorderedMap;
 use crate::collections::{append, Vector};
+use crate::IntoStorageKey;
 
 /// TreeMap based on AVL-tree
 ///
@@ -46,11 +47,15 @@ where
     K: Ord + Clone + BorshSerialize + BorshDeserialize,
     V: BorshSerialize + BorshDeserialize,
 {
-    pub fn new(id: Vec<u8>) -> Self {
+    pub fn new<S>(prefix: S) -> Self
+    where
+        S: IntoStorageKey,
+    {
+        let prefix = prefix.into_storage_key();
         Self {
             root: 0,
-            val: UnorderedMap::new(append(&id, b'v')),
-            tree: Vector::new(append(&id, b'n')),
+            val: UnorderedMap::new(append(&prefix, b'v')),
+            tree: Vector::new(append(&prefix, b'n')),
         }
     }
 
@@ -759,7 +764,7 @@ mod tests {
     fn test_empty() {
         test_env::setup();
 
-        let map: LegacyTreeMap<u8, u8> = LegacyTreeMap::new(vec![b't']);
+        let map: LegacyTreeMap<u8, u8> = LegacyTreeMap::new(b't');
         assert_eq!(map.len(), 0);
         assert_eq!(height(&map), 0);
         assert_eq!(map.get(&42), None);
@@ -909,7 +914,7 @@ mod tests {
         let n: u64 = 30;
         let vec = random(n);
 
-        let mut map: LegacyTreeMap<u32, u32> = LegacyTreeMap::new(vec![b't']);
+        let mut map: LegacyTreeMap<u32, u32> = LegacyTreeMap::new(b't');
         for x in vec.iter().rev() {
             map.insert(x, &1);
         }
@@ -925,7 +930,7 @@ mod tests {
         let n: u64 = 30;
         let vec = random(n);
 
-        let mut map: LegacyTreeMap<u32, u32> = LegacyTreeMap::new(vec![b't']);
+        let mut map: LegacyTreeMap<u32, u32> = LegacyTreeMap::new(b't');
         for x in vec.iter().rev() {
             map.insert(x, &1);
         }