+++ near-sdk/src/collections/unordered_set.rs
@@ -1,7 +1,7 @@
 //! A set implemented on a trie. Unlike `std::collections::HashSet` the elements in this set are not
 //! hashed but are instead serialized.
 use crate::collections::{append, append_slice, Vector};
-use crate::env;
+use crate::{env, IntoStorageKey};
 use borsh::{BorshDeserialize, BorshSerialize};
 use std::mem::size_of;
 
@@ -27,9 +27,13 @@ impl<T> UnorderedSet<T> {
     }
 
     /// Create new map with zero elements. Use `id` as a unique identifier.
-    pub fn new(id: Vec<u8>) -> Self {
-        let element_index_prefix = append(&id, b'i');
-        let elements_prefix = append(&id, b'e');
+    pub fn new<S>(prefix: S) -> Self
+    where
+        S: IntoStorageKey,
+    {
+        let prefix = prefix.into_storage_key();
+        let element_index_prefix = append(&prefix, b'i');
+        let elements_prefix = append(&prefix, b'e');
 
         Self { element_index_prefix, elements: Vector::new(elements_prefix) }
     }
@@ -179,7 +183,7 @@ mod tests {
     #[test]
     pub fn test_insert() {
         test_env::setup();
-        let mut set = UnorderedSet::new(b"s".to_vec());
+        let mut set = UnorderedSet::new(b"s");
         let mut rng = rand_xorshift::XorShiftRng::seed_from_u64(0);
         for _ in 0..500 {
             let key = rng.gen::<u64>();
@@ -190,7 +194,7 @@ mod tests {
     #[test]
     pub fn test_insert_remove() {
         test_env::setup();
-        let mut set = UnorderedSet::new(b"s".to_vec());
+        let mut set = UnorderedSet::new(b"s");
         let mut rng = rand_xorshift::XorShiftRng::seed_from_u64(1);
         let mut keys = vec![];
         for _ in 0..100 {
@@ -207,7 +211,7 @@ mod tests {
     #[test]
     pub fn test_remove_last_reinsert() {
         test_env::setup();
-        let mut set = UnorderedSet::new(b"s".to_vec());
+        let mut set = UnorderedSet::new(b"s");
         let key1 = 1u64;
         set.insert(&key1);
         let key2 = 2u64;
@@ -223,7 +227,7 @@ mod tests {
     #[test]
     pub fn test_insert_override_remove() {
         test_env::setup();
-        let mut set = UnorderedSet::new(b"s".to_vec());
+        let mut set = UnorderedSet::new(b"s");
         let mut rng = rand_xorshift::XorShiftRng::seed_from_u64(2);
         let mut keys = vec![];
         for _ in 0..100 {
@@ -244,7 +248,7 @@ mod tests {
     #[test]
     pub fn test_contains_non_existent() {
         test_env::setup();
-        let mut set = UnorderedSet::new(b"s".to_vec());
+        let mut set = UnorderedSet::new(b"s");
         let mut rng = rand_xorshift::XorShiftRng::seed_from_u64(3);
         let mut set_tmp = HashSet::new();
         for _ in 0..500 {
@@ -261,7 +265,7 @@ mod tests {
     #[test]
     pub fn test_to_vec() {
         test_env::setup();
-        let mut set = UnorderedSet::new(b"s".to_vec());
+        let mut set = UnorderedSet::new(b"s");
         let mut rng = rand_xorshift::XorShiftRng::seed_from_u64(4);
         let mut keys = HashSet::new();
         for _ in 0..500 {
@@ -276,7 +280,7 @@ mod tests {
     #[test]
     pub fn test_clear() {
         test_env::setup();
-        let mut set = UnorderedSet::new(b"s".to_vec());
+        let mut set = UnorderedSet::new(b"s");
         let mut rng = rand_xorshift::XorShiftRng::seed_from_u64(5);
         for _ in 0..10 {
             for _ in 0..=(rng.gen::<u64>() % 20 + 1) {
@@ -292,7 +296,7 @@ mod tests {
     #[test]
     pub fn test_iter() {
         test_env::setup();
-        let mut set = UnorderedSet::new(b"s".to_vec());
+        let mut set = UnorderedSet::new(b"s");
         let mut rng = rand_xorshift::XorShiftRng::seed_from_u64(4);
         let mut keys = HashSet::new();
         for _ in 0..500 {
@@ -307,7 +311,7 @@ mod tests {
     #[test]
     pub fn test_extend() {
         test_env::setup();
-        let mut set = UnorderedSet::new(b"s".to_vec());
+        let mut set = UnorderedSet::new(b"s");
         let mut rng = rand_xorshift::XorShiftRng::seed_from_u64(4);
         let mut keys = HashSet::new();
         for _ in 0..100 {